requirement.txt
discord.py
yfinance
python-dotenv
requests
feedparser